+++ database/migrations/2023_04_03_050610_create_detail_receipts_table.php
@@ -15,7 +15,7 @@ class CreateDetailReceiptsTable extends Migration
     {
         Schema::create('detail_receipts', function (Blueprint $table) {
             $table->id();
-            $table->decimal('price'); //don gia
+            $table->double('price'); //don gia
             $table->unsignedInteger('quantity');
             $table->bigInteger('product_id')->unsigned()->nullable();
             $table->bigInteger('warehouse_id')->unsigned()->nullable();
